backend/requirements.txt
python-jose
bcrypt==4.1.3
passlib==1.7.4
pydantic-settings
fastapi
uvicorn
python-multipart
sqlalchemy>=2.0.36
email-validator
alembic
aiofiles==23.2.1
pytest
fastapi-cache2==0.2.1
redis==5.0.1